fix(web): avoid daemon downgrade guidance
