+++ ARCHITECTURE.md
@@ -107,7 +107,7 @@ Lets any machine join a server without cloning the repo: `npx @fancyboi999/open-
 - `views/Chat.tsx` / `AgentActivity.tsx` — Main group-chat view plus the shared per-message Activity disclosure. Both channel rows and thread rows consume the same channel-scoped `agent:reply` events, persist/merge `message:updated` tails, render actual public messages only after message API persistence, and render no-message receipts without body/actions. Live placeholders opt into automatic expansion; persisted messages and receipts default to collapsed even if more events arrive. Header run state is derived from the message snapshot rather than the agent's workspace-global presence. The fourth grid column opens only for a thread or member profile; Agent Profile Activity remains the global diagnostic surface.
 - `views/Composer.tsx` — Shared message composer for channels, DMs, and threads: text + attachment upload (button / paste / drag-drop, with progress) + @mention autocomplete + send. `allowAsTask` gates the As Task toggle (channels/DMs only; a thread reply is never a task). Renders a **reachability/wake hint** above the input for the agents a message will reach — the DM peer (`dmAgent` prop) + any @-mentioned in the draft — when their machine is offline (copy: the message is saved and handled automatically once the machine reconnects, via reconnect catch-up), or, for a DM, when the sole peer is sleeping. Centralized here (replacing the former DM-only `topSlot` hint) so channels/DMs/threads share one path.
 - `toast.tsx` — Lightweight global toast notifications (`ToastProvider` / `useToast()`, mirrors `ConfirmModal`'s context shape, no third-party lib; mounted in `main.tsx`). One-shot operation feedback where an inline banner doesn't fit — e.g. start/restart an agent whose machine is offline (503), or create an agent that couldn't start (`started:false`).
-- `alerts.tsx` — **System alert center**: `useSystemAlerts()` derives *standing* system warnings from live store state (an online machine whose `daemonVersion` ≠ `latestDaemonVersion` → outdated daemon; an offline machine still hosting agents → those agents can't run), and `NotificationCenter` renders them in a fixed, rail-anchored popover. The rail's warning button (above Settings in `Layout.tsx`) shows only when there are undismissed alerts. Distinct from `toast.tsx`'s one-shot feedback — alerts recompute from state and are individually dismissable (session-scoped). `latestDaemonVersion` is read from `packages/daemon/package.json` by the `/api/servers/:id/machines` route (`routes-api/servers.ts`).
+- `alerts.tsx` / `machineUi.ts` — **System alert center + daemon version guidance**: `useSystemAlerts()` warns only when an online machine's valid semantic `daemonVersion` is lower than `latestDaemonVersion`; the same pure comparator gates Computers' update action, so equal, newer, missing, and development versions never suggest a downgrade. `NotificationCenter` renders standing warnings in a rail-anchored popover; `latestDaemonVersion` comes from `packages/daemon/package.json` through the machines route.
 - `views/Members.tsx` / `views/ProjectDirectoryPicker.tsx` — Members directory: the sidebar groups agents by machine + a humans section; the default roster lists agents and humans as two labelled sections of clickable cards, each navigating into that member's profile. AgentProfile seven-tab panel (overview / activity / workspace / permissions / apps / DM / reminders) + HumanProfile + create-agent modal. Create and inactive edit retain manual project paths and add a portal-based remote directory picker with independent roots/discovery loading, lazy child navigation, stale-request guards, and desktop/390px layouts; absolute paths travel only in POST bodies.
 - `views/ChatSidebar.tsx` — channel/DM sidebar shared by the Chat and Saved views (channel list, DM list, unread badges; hosts `LiveAgentBar` at its bottom).
 - `views/ConnectComputerWizard.tsx` — multi-step connect-a-computer modal behind all three entry points (`onboard` auto-show / Computers `add` / offline-machine `reconnect` with key rotation).
